Generating a consistent story to BotServices between LUIS and QnAMaker
